src/main/java/com/imooc/demo/service/Iml/CategoryServiceIml.java
package com.imooc.demo.service.Iml;


import com.imooc.demo.DataObject.ProductCategory;
import com.imooc.demo.Repository.ProductCategoryRepository;
import com.imooc.demo.service.CategoryService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class CategoryServiceIml implements CategoryService {

    @Autowired
    private ProductCategoryRepository repository;

    @Override
    public ProductCategory findOne(Integer categoryId) {
        return repository.findOne(categoryId);
    }

    @Override
    public List<ProductCategory> findAll() {
        return repository.findAll();
    }

    @Override
    public ProductCategory save(ProductCategory productCategory) {
        return repository.save(productCategory);
    }

    @Override
    public List<ProductCategory> findByCategoryTypeIn(List<Integer> categoryTypeList){
        return repository.findByCategoryTypeIn(categoryTypeList);
    };
}
